* The function no longer crashes * but the prompts need to be optimized.
