+++ Dockerfile_mcp
@@ -16,4 +16,8 @@ RUN pip install pipenv
 
 RUN pipenv lock && pipenv install --system --deploy --ignore-pipfile
 
+RUN crawl4ai-setup
+
+RUN crawl4ai-doctor
+
 CMD [ "python" , "main.py"]
